Post List: Fix the response after saving a Quick Edit. (#21718)
Post List: Fix the response after saving a Quick Edit.

The customizations that to the post list screen were being made on hooks
that only ran as the page/screen loaded. If a quick edit was made, the
table row returned in the result didn't include the necessary
customizations and would break the table.

Co-authored-by: Dean Sas <dean.sas@automattic.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1480367561